# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 동인고등학교 업무 자동화 대시보드 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  parentId: number | null;
  createdAt: string;
  children?: Folder[];
  projectCount?: number;
}

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  parentId?: number | null;
}

export interface FolderUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  parentId?: number | null;
}

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  creator: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  folderName?: string | null;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minLength 1 */
  url: string;
  /** @minLength 1 */
  creator: string;
  /** @nullable */
  folderId?: number | null;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  url?: string;
  creator?: string;
  /** @nullable */
  folderId?: number | null;
}

export interface DashboardStats {
  totalProjects: number;
  totalFolders: number;
  recentCount: number;
}

export type ListProjectsParams = {
/**
 * @nullable
 */
folderId?: number | null;
};