# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 동인고등학교 업무 자동화 대시보드 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: folders
    description: Folder operations
  - name: projects
    description: Project operations
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List all folders as a tree
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Created folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"

  /folders/{id}:
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Update a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderUpdate"
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects, optionally filtered by folder
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /projects/recent:
    get:
      operationId: getRecentProjects
      tags: [projects]
      summary: Get recently added projects
      responses:
        "200":
          description: Recent projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Folder:
      type: object
      required: [id, name, parentId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        children:
          type: array
          items:
            $ref: "#/components/schemas/Folder"
        projectCount:
          type: integer

    FolderInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: ["integer", "null"]

    FolderUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: ["integer", "null"]

    Project:
      type: object
      required: [id, title, url, creator, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        creator:
          type: string
        folderId:
          type: ["integer", "null"]
        folderName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    ProjectInput:
      type: object
      required: [title, url, creator]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        url:
          type: string
          minLength: 1
        creator:
          type: string
          minLength: 1
        folderId:
          type: ["integer", "null"]

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        url:
          type: string
        creator:
          type: string
        folderId:
          type: ["integer", "null"]

    DashboardStats:
      type: object
      required: [totalProjects, totalFolders, recentCount]
      properties:
        totalProjects:
          type: integer
        totalFolders:
          type: integer
        recentCount:
          type: integer