# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 동인고등학교 업무 자동화 대시보드 API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  url?: string;
  creator?: string;
  /** @nullable */
  folderId?: number | null;
}